Move utilities to separate folder
